backend/pyproject.toml
[project]
name = "my-tasklist-backend"
version = "0.2.0"
description = "FastAPI backend for the tasklist app"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "pydantic-settings==2.5.2",
    "SQLAlchemy==2.0.35",
    "PyMySQL==1.1.0",
    "python-dotenv==1.0.0",
    "cryptography",
    "requests==2.31.0",
    "werkzeug==3.0.4",
    "itsdangerous==2.2.0",
]
